Major update to how db is initialized 1. maven drops/recreates the db schema any time the tests are run (it is bound to "test" goal) 2. db machine name can now be configurable in ~/.m2/settings.xml: <profiles>     <profile>       <id>db-config</id>       <properties>         <dbServer>localhost</dbServer>       </properties>     </profile>   </profiles>
  <activeProfiles>
    <activeProfile>db-config</activeProfile>
  </activeProfiles>

otherwise, it defaults to "localhost" (specified in core/pom.xml default profile).
3. Still using the db as 'junit-<username>', but that's being set in the core/pom.xml and no longer in DBTestSuite
In fact, all the hibernate config stuff is done in the core/pom now, using an ANT task that copies the template file (so the individual's hibernate.cfg.xml is no longer checked in), replacing the server and dbname and runnign the hibernatetool to recreate the db.
4. DBTestSuite now only deletes data from all tables, and then loads the Currencies and the sub-account-types.
5. Annotations should be producing valid useable schema now




git-svn-id: http://source.marketcetera.org/root/trunk@320 501a4a96-914b-0410-8454-f5988a4ca596