Fix for bug #477 - trades recorded after midnight 'today' weren't showing up in Positions list for 'today', and weren't counted in 'number of trades today' in welcome dashboard.
Also modified messages_log controller - it's now essentially only for display of unparseable messages.
The !TradeRecorder will still put an entry in the case it can't parse an incoming execution report into the table, so this is the way to display them.
Got rid of the screens/workflow to display recorded execution reports altogether



git-svn-id: http://source.marketcetera.org/root/trunk@8995 501a4a96-914b-0410-8454-f5988a4ca596