Fix for EG-341, reviewed as ENG-190, changes as follows:
This change addresses the issue raised in ENG-341.  When trying to register a ruby script through Photon, the registration process failed and the following stack-trace 
appeared in the log (partially elided):
!MESSAGE Unhandled event loop exception
!STACK 0
java.lang.VerifyError: (class: org/marketcetera/strategy/StrategyModule, method: cancelAllMarketDataRequests signature:
()V) Attempt to split long or double on the stack

The source of the problem is in JRuby which has a bug in its bindings of scalar types > 32 bits (includes doubles and floats, but not BigDecimals or int) if the types in 
question are used in the method signature of a method which was originally declared in an interface (what?!?! are you kidding?).  The problem is fixed in a later version of 
JRuby, but we cannot upgrade from 1.1.3 without some significant changes in how Ruby scripts are invoked.  We cannot make those changes now, so, instead, the fix I 
implemented was to remove the use of Java "long" types for request identifiers.  Instead, I used "int" types.  The effective difference is that users are limited to making 
2,147,483,647 simultaneous data requests across all strategies in a single JVM session instead of 9,223,372,036,854,775,807.  I strongly suspect that this will be adequate.



git-svn-id: http://source.marketcetera.org/root/trunk@10192 501a4a96-914b-0410-8454-f5988a4ca596