Author: tlerios
Reviewer: ytalmor

Crucible: ENG-196

JIRA: EG-316

Comments:

+ Revised the generation of ACKs and rejects generated by the ORS.

 - ACK changes (regular execution reports) did not change except in
   that a dummy OrderID as well as the standard heater/trailer fields
   (with dummy values) were added to make them valid ERs. No
   functional impact.

 - Rejects changed substantially in that many more fields are now
   filled-in, and the resulting reply message is a valid FIX
   message. Accordingly, Photon demonstrates a better behavior, and
   the FIX Message View for these rejects is much more informative.

 - In both cases, fixed the generation of text fields to replace SOH's
   with |'s, or else the FIX message may not convert to string
   correctly (and thus marshalling to Photon will fail).

+ Added utilities to create empty OrderCancelRejects and
ExecutionReports in core.

+ Moved logging of heartbeats to ERROR to hide the regular messages
warning that heartbeats could not be sent to the client.

+ Moved the last legacy message from the ors message file to a
separate file for the mbeans package.

+ Minor improvements to comments and log messages.

Note: the changes to RequestHandler are more limited than this diff
would lead you to believe: code moved from the bottom of the file to
the top, and rather than showing a true diff of that code, there's a
large block of additions followed by a large block of removals. But
the code itself did not change dramatically. That was also the last
remaining legacy code in the ors package (there is still plenty in
subpackages, but it plays a less critical role).



git-svn-id: http://source.marketcetera.org/root/trunk@10198 501a4a96-914b-0410-8454-f5988a4ca596